✨ feat: google map 객체 전역 state로 설정
