+++ src/app/(Main)/Map/page.tsx
@@ -1,21 +1,18 @@
 'use client';
 
 import { useEffect } from 'react';
+import { useSetRecoilState } from 'recoil';
+import { mapState } from '@/recoil/atoms';
+import { createMap } from '@/service/map';
 
 function Map() {
+  const setMap = useSetRecoilState(mapState);
+
   // 아마 차후에 추상팩토리같은 디자인 패턴을 적용해야 하지않을까 싶네요.
-  const addGoogleMap = (center: google.maps.LatLngLiteral): google.maps.Map => new google.maps.Map(document.getElementById('map') as HTMLElement, {
-    center,
-    zoom: 15,
-  });
   const style = { width: '100vw', height: '100vh' };
   useEffect(() => {
-    const center: google.maps.LatLngLiteral = {
-      lat: 37.3595316,
-      lng: 127.1052133,
-    };
-    addGoogleMap(center);
-  }, []);
+    setMap(createMap());
+  }, [setMap]);
   return (
     <>
       <div id="map" style={style} />
