+++ src/components/LeftNav.tsx
@@ -2,16 +2,16 @@ import styled, { createGlobalStyle } from 'styled-components';
 // import MapNode from "@/service/MapObject/MapNode";
 import MapNode from '@/service/MapObject/MapNode';
 import { useRef, useState, useEffect } from 'react';
-import DraggablePieChart from '@/ui/DraggablePieChart';
 import MapNodeCard from '@/ui/MapNodeCard';
 import searchNearbyPlace from '@/service/search';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import {
-  searchAddressState, searchClickState, useSsrComplectedState, ssrCompletedState,
+  searchAddressState, searchClickState, useSsrComplectedState, ssrCompletedState, mapState,
 } from '@/recoil/atoms';
 
 function LeftNavContainer({ isLeftNavOpen }: { isLeftNavOpen: boolean }) {
   const [searchMapNodes, setSearchMapNodes] = useState<MapNode[]>([]);
+  const map = useRecoilValue(mapState);
   const searchAddress = useRecoilValue(searchAddressState);
   const isSearchClick = useRecoilValue(searchClickState);
   const checkSSREnd = useSsrComplectedState();
@@ -21,7 +21,7 @@ function LeftNavContainer({ isLeftNavOpen }: { isLeftNavOpen: boolean }) {
       checkSSREnd();
     } else {
       const handleSearchClick = async () => {
-        const sortedMapNodes: MapNode[] = await searchNearbyPlace(searchAddress);
+        const sortedMapNodes: MapNode[] = await searchNearbyPlace(searchAddress, map);
         setSearchMapNodes(sortedMapNodes);
       };
       handleSearchClick();
@@ -32,9 +32,6 @@ function LeftNavContainer({ isLeftNavOpen }: { isLeftNavOpen: boolean }) {
     <>
       <GlobalStyle />
       <LeftNavStyled isLeftNavOpen={isLeftNavOpen}>
-        {/* <DraggablePieChart /> */}
-        <TitleStyled>test</TitleStyled>
-
         {searchMapNodes.map((node, i) => (
           <MapNodeCard key={node.id} index={i + 1} node={node} />
         ))}
@@ -55,12 +52,12 @@ const TitleStyled = styled.div`
 
 const LeftNavStyled = styled.div<{ isLeftNavOpen: boolean }>`
   position: fixed;
-  top: 100px;
-  left: ${(props) => (props.isLeftNavOpen ? '0px' : '-300px')}; /* 변경된 부분 */
+  top: 120px;
+  left: ${(props) => (props.isLeftNavOpen ? '-20px' : '-300px')}; /* 변경된 부분 */
   display: flex;
   flex-direction: column;
   align-items: center;
-  background-color: #ffb9b4;
+  //background-color: #ffb9b4;
   width: 300px;
   height: calc(100% - 100px);
   overflow-y: auto;
@@ -73,4 +70,7 @@ const GlobalStyle = createGlobalStyle`
     src: url(//fonts.gstatic.com/ea/nanumgothic/v5/NanumGothic-Regular.eot);
     font-family: 'Nanum Gothic', serif;
   }
+  ::-webkit-scrollbar {
+  display: none;
+}
 `;
